Timing data analysis completed after 1st opt
After some modifications, the new analysis on the usage percentage
has been conducted. The modifications introducted are
In the Constructor
-- The barrel shifters has been replaced with a couple of shifters of 1 clock
cycle. This has reduced the cost but added 5 clock cycles of latency in the constructor.
(2 are due to the HoW component, 3 are due to the LR_SH2 component).
In the Extimator
-- The Barrel shifters has been replaced with two shifters cascaded. This has added 2 clk cycles of latency.
-- The 11 bit multiplier has been replaced with a 6 bit multiplier which works for 4 clock cycles. This
Has increased the usage percentage by a factor of 4 but it introduced 4 clk cycles of latency.
-- The 17 bit Three-input adder has been replaced with a 19 bit Two-input adder. This has doubled the usage percentage
but added 1 clock cycle of latency.
-- The Pel_adder (a 4-input adder) has been replaced with two 2-input adder and a 2-input adder cascaded.
This has added 1 clock cycle of latency.
In conclusion, the usage percentage has been incremented and the cost has been slightly reduced,
but the latency has slightly increased. This is not a problem since we are still in the
constraint of the 500MHz.